+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - General source code linting
 - Use IntoIterator trait as interface in methods that require an iterator
 - Changed iterator names (C-ITER-TY API guidelines)
+- Changed linear system solvers names (C-ITER-TY API guidelines)
 ## Fixed
 - Subtraction between a real number and a polynomial